+++ python/eyes_robotframework/src/EyesLibrary/keywords/config.py
@@ -1,15 +1,18 @@
 from __future__ import absolute_import, unicode_literals
 
+from copy import deepcopy
 from datetime import datetime
 from typing import Dict, Optional, Text, Union
 
+from robot.api import logger as robot_logger
 from robot.api.deco import keyword
 from six import string_types as basestring
 
 from applitools.common import deprecated
-from applitools.selenium import BatchInfo
+from applitools.selenium import BatchInfo, Eyes
 
 from ..base import LibraryComponent
+from ..config_parser import SelectedRunner
 
 
 class ConfigurationKeywords(LibraryComponent):
@@ -77,3 +80,40 @@ class ConfigurationKeywords(LibraryComponent):
     def eyes_configure_add_property(self, name, value):
         # type: (Text, Text) -> None
         self.ctx.configure.add_property(name, value)
+
+    @keyword("Eyes Set NMG Capabilities", types=(dict,))
+    def eyes_set_nmg_capabilities(self, capabilities):
+        # type:  (Dict) -> None
+        """Inplace update desired capabilities with NMG required parameters.
+
+            | =Arguments=                  | =Description=                                                                              |
+            | Capabilities                 | The capabilities for mobile device
+
+        *Example:*
+            | Eyes Set NMG Capabilities | ${DESIRED_CAPS}  |
+        """
+        if self.selected_runner is not SelectedRunner.native_mobile_grid:
+            robot_logger.warn(
+                "Add NMG capabilities make sense use only with `native_mobile_grid` runner. Skipping..."
+            )
+            return
+        config = self.get_configuration()
+        Eyes.set_nmg_capabilities(
+            capabilities, config.api_key, config.server_url, config.proxy
+        )
+
+    @keyword("Eyes Create NMG Capabilities")
+    def eyes_create_nmg_capabilities(self, **kwargs):
+        # type:  (**Dict)-> Dict
+        """Creates new capabilities with NMG related fields.
+
+        Capabilities of appium server, Android and iOS,
+        Please check https://github.com/appium/appium/blob/master/docs/en/writing-running-appium/server-args.md
+
+        Examples:
+        | ${CAPS}=| Eyes Create NMG Capabilities | platformName=iOS      | platformVersion=7.0            | deviceName='iPhone Simulator'           | app=your.app                         |
+        | ${CAPS}=| Eyes Create NMG Capabilities | platformName=Android | platformVersion=4.2.2 | deviceName=192.168.56.101:5555 | app=${CURDIR}/demoapp/OrangeDemoApp.apk | appPackage=com.netease.qa.orangedemo | appActivity=MainActivity |
+        """
+        capabilities = deepcopy(kwargs)
+        self.eyes_set_nmg_capabilities(capabilities)
+        return capabilities